refactor(deskwork): extract @deskwork/cli; collapse plugin to a thin shell
Second commit toward separating headless skills from in-tree code.
The plugin now contains only SKILL.md files plus a small bin/ wrapper
that delegates to the @deskwork/cli npm package. CLI logic lives in its
own workspace and exposes a single dispatch entry, `deskwork`.

Architecture before:
  plugins/deskwork/
    bin/deskwork-add.ts            (and 11 more — top-level scripts)
    package.json                    (tsx, vitest, typescript devDeps)
    skills/<name>/SKILL.md         (invoked bin/X.ts)
    test/*-integration.test.ts     (spawned bin/X.ts)
    tsconfig.json, vitest.config.ts

After:
  packages/cli/
    src/cli.ts                      (single dispatch entry — `bin: deskwork`)
    src/commands/<name>.ts          (export run(argv) — refactored from bin/X.ts)
    test/*-integration.test.ts      (spawn node_modules/.bin/deskwork)
    tsconfig.json, vitest.config.ts, package.json

  plugins/deskwork/
    bin/deskwork                    (bash wrapper — workspace-link first,
                                     `npx -y @deskwork/cli@latest` fallback)
    package.json                    (metadata only)
    skills/<name>/SKILL.md          (invoke `deskwork <subcommand>`)

What this enables:
- The plugin updates independently of the CLI. Skill prose changes
  don't require a new CLI release; CLI bug fixes don't require a
  plugin update.
- Headless skills don't carry tsx/vitest/typescript weight in the
  plugin tree — those are dev deps of @deskwork/cli only.
- After publishing @deskwork/cli to npm, the plugin install becomes
  a few KB of markdown + a 30-line shell wrapper that fetches the
  CLI from npm via npx.

Path-injection convenience: cli.ts inserts process.cwd() as the
project-root positional when the user doesn't pass one. So
`deskwork add 'My Post'` from /work/myproject is equivalent to
`deskwork add /work/myproject 'My Post'`. Tests that pass an
explicit project-root keep working — the heuristic only triggers
when argv[0] doesn't look path-like (no `/`, `./`, `../`, or `.`).

Verified:
- 126/126 @deskwork/core tests pass
- 27/27 @deskwork/cli tests pass (3 integration files spawn the
  workspace `deskwork` binary; same coverage as before the move)
- Both packages typecheck clean
- claude plugin validate plugins/deskwork passes
- claude --plugin-dir plugins/deskwork lists all 12 skills
- bin/deskwork wrapper resolves the workspace-linked binary and
  dispatches correctly

Next commits: build @deskwork/studio (Hono server + UI), add the
deskwork-studio plugin shell, end-to-end dogfood.
