+++ .git-commit-msg.tmp
@@ -1,41 +1,57 @@
-refactor(deskwork): extract @deskwork/core npm package
+refactor(deskwork): extract @deskwork/cli; collapse plugin to a thin shell
 
-First commit toward severing the plugin from in-tree code. Per the plan
-in /Users/orion/.claude/plans/i-would-like-to-wiggly-hennessy.md, the
-target is a thin SKILL-only plugin pointing at npm-published packages
-(@deskwork/cli, @deskwork/studio later) so:
-  - Headless skills don't carry studio weight
-  - Studio doesn't depend on Astro
-  - Plugin updates decouple from npm-package updates
+Second commit toward separating headless skills from in-tree code.
+The plugin now contains only SKILL.md files plus a small bin/ wrapper
+that delegates to the @deskwork/cli npm package. CLI logic lives in its
+own workspace and exposes a single dispatch entry, `deskwork`.
 
-This commit moves the lib code out of the plugin tree into its own
-workspace package. Behavior unchanged. The bin/ scripts and
-integration tests now resolve `@deskwork/core/<subpath>` through the
-npm-workspace symlink instead of relative `../lib/X.ts` imports.
+Architecture before:
+  plugins/deskwork/
+    bin/deskwork-add.ts            (and 11 more — top-level scripts)
+    package.json                    (tsx, vitest, typescript devDeps)
+    skills/<name>/SKILL.md         (invoked bin/X.ts)
+    test/*-integration.test.ts     (spawned bin/X.ts)
+    tsconfig.json, vitest.config.ts
 
-Changes:
-- Add packages/core/ workspace with src/, test/, package.json, tsconfig,
-  vitest config. Subpath exports cover all internal modules.
-- Move plugins/deskwork/lib/* → packages/core/src/* (git mv preserves
-  history)
-- Move 9 lib tests → packages/core/test/* with `../src/X.ts` imports
-- Update plugins/deskwork/package.json: drop yaml/unified/remark/
-  rehype-stringify (now @deskwork/core deps); declare @deskwork/core: *
-- Update plugins/deskwork/bin/* and 3 integration tests to import from
-  @deskwork/core/<subpath>
-- Update root package.json workspaces: ["packages/*", "plugins/*"]
-- Update review-lifecycle-integration.test.ts inline tsx scripts to
-  use @deskwork/core/review/pipeline + @deskwork/core/config (no more
-  pluginRoot/lib/...)
+After:
+  packages/cli/
+    src/cli.ts                      (single dispatch entry — `bin: deskwork`)
+    src/commands/<name>.ts          (export run(argv) — refactored from bin/X.ts)
+    test/*-integration.test.ts      (spawn node_modules/.bin/deskwork)
+    tsconfig.json, vitest.config.ts, package.json
+
+  plugins/deskwork/
+    bin/deskwork                    (bash wrapper — workspace-link first,
+                                     `npx -y @deskwork/cli@latest` fallback)
+    package.json                    (metadata only)
+    skills/<name>/SKILL.md          (invoke `deskwork <subcommand>`)
+
+What this enables:
+- The plugin updates independently of the CLI. Skill prose changes
+  don't require a new CLI release; CLI bug fixes don't require a
+  plugin update.
+- Headless skills don't carry tsx/vitest/typescript weight in the
+  plugin tree — those are dev deps of @deskwork/cli only.
+- After publishing @deskwork/cli to npm, the plugin install becomes
+  a few KB of markdown + a 30-line shell wrapper that fetches the
+  CLI from npm via npx.
+
+Path-injection convenience: cli.ts inserts process.cwd() as the
+project-root positional when the user doesn't pass one. So
+`deskwork add 'My Post'` from /work/myproject is equivalent to
+`deskwork add /work/myproject 'My Post'`. Tests that pass an
+explicit project-root keep working — the heuristic only triggers
+when argv[0] doesn't look path-like (no `/`, `./`, `../`, or `.`).
 
 Verified:
 - 126/126 @deskwork/core tests pass
-- 27/27 @deskwork/plugin tests pass (3 integration files spawning real
-  bin scripts via tsx)
-- npm --workspaces run typecheck clean
+- 27/27 @deskwork/cli tests pass (3 integration files spawn the
+  workspace `deskwork` binary; same coverage as before the move)
+- Both packages typecheck clean
 - claude plugin validate plugins/deskwork passes
-- claude --plugin-dir lists all 12 skills
+- claude --plugin-dir plugins/deskwork lists all 12 skills
+- bin/deskwork wrapper resolves the workspace-linked binary and
+  dispatches correctly
 
-Next commits per the plan: extract @deskwork/cli (bin/ → packages/cli/),
-build @deskwork/studio (Hono server + UI), add deskwork-studio plugin
-shell.
+Next commits: build @deskwork/studio (Hono server + UI), add the
+deskwork-studio plugin shell, end-to-end dogfood.
