+++ packages/cli/test/review-lifecycle-integration.test.ts
@@ -19,8 +19,8 @@ import { dirname, join, resolve } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
 const testDir = dirname(fileURLToPath(import.meta.url));
-const pluginRoot = resolve(testDir, '..');
-const bin = (name: string) => join(pluginRoot, 'bin', `${name}.ts`);
+const workspaceRoot = resolve(testDir, '../../..');
+const deskworkBin = join(workspaceRoot, 'node_modules/.bin/deskwork');
 
 interface RunResult {
   code: number;
@@ -30,7 +30,8 @@ interface RunResult {
 }
 
 function run(script: string, args: string[]): RunResult {
-  const r = spawnSync(bin(script), args, { encoding: 'utf-8' });
+  const subcommand = script.replace(/^deskwork-/, '');
+  const r = spawnSync(deskworkBin, [subcommand, ...args], { encoding: 'utf-8' });
   const stdout = r.stdout ?? '';
   let json: unknown;
   try {
