--- plugins/deskwork/vitest.config.ts
@@ -1,16 +1,0 @@
-import { defineConfig } from 'vitest/config';
-import { fileURLToPath } from 'node:url';
-import { dirname, resolve } from 'node:path';
-
-const __dirname = dirname(fileURLToPath(import.meta.url));
-
-export default defineConfig({
-  test: {
-    include: ['test/**/*.test.ts'],
-  },
-  resolve: {
-    alias: {
-      '@': resolve(__dirname, '.'),
-    },
-  },
-});
